Add getFirstEditionFromYear function + cleanup
Add periodical/helper.ts
+ Add getFirstEditionFromYear 
+ Move GroupListItem
