+++ src/components/material/MaterialHeader.tsx
@@ -19,9 +19,9 @@ import { Cover } from "../cover/cover";
 import MaterialAvailabilityText from "./MaterialAvailabilityText/MaterialAvailabilityText";
 import MaterialHeaderText from "./MaterialHeaderText";
 import MaterialButtons from "./material-buttons/MaterialButtons";
-import MaterialPeriodical from "./MaterialPeriodical";
+import MaterialPeriodical from "./periodical/MaterialPeriodical";
 import { Manifestation, Work } from "../../core/utils/types/entities";
-import { GroupListItem } from "./MaterialPeriodicalSelect";
+import { GroupListItem } from "./periodical/helper";
 
 interface MaterialHeaderProps {
   wid: WorkId;
